rendu3/requirements.txt
nltk==3.8.1
spacy>=3.4.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
pandas>=1.3.0
pyyaml>=6.0